backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "langchain>=0.3.25",
    "langchain-openai>=0.3.17",
    "langgraph>=0.4.5",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
]
